ajout client

+++ composition-bouquets/backend/src/main/java/com/atelierchantdefleur/bouquetcomposer/model/domain/MateriauDTO.java
@@ -18,6 +18,6 @@ public class MateriauDTO {
     @NotEmpty
     private String nom;
     @Digits(integer = 6, message = "Format du prix unitaire invalide", fraction = 0)
-    private Integer prixUnitaire;
+    private Double prixUnitaire;
     private FournisseurDTO fournisseurDTO;
 }
